poc/backend/requirements.txt
fastapi
uvicorn
pydantic
langchain
langgraph
langchain-google-genai
python-dotenv
